Import directive definitions from custom prototype (#92)
Using a custom prototype with the skip_prototype option would fail the
federated schema composition because absinthe will not add required
directive and type definitions from the prototype. This changes the way we
use a custom prototype; instead of providing a skip_prototype boolean and
a module attribute for the prototype, we provide the prototype_schema
option when we `use Absinthe.Federation.Schema`. This way, the definitions
are imported as necessary.

This change also unlocks the @composeDirective support in the
federation_compatibility checks.

The breaking change from 0.4.2 will be added to the changelog.